feat(DeleteConfirmationModal)Add isSubmitting state

+++ src/components/templateDepend/templateDetails/commentsSection/deleteComment/DeleteComment.jsx
@@ -1,12 +1,15 @@
+import { useEffect, useState } from 'react';
 import { useUser } from '../../../../../contexts';
 import useFetch from '../../../../../hooks/useFetch';
+import DeleteConfirmationModalPortal from '../../../../../portals/DeleteConfirmationModalProtal';
 import styles from './DeleteComment.module.css';
 import DeleteConfirmationModal from './DeleteConfirmationModal';
 
 export default function DeleteComment({ comment, updateCommentHandler }) {
 	const { user } = useUser();
-
-	const { error, refetch } = useFetch(
+	const [showConfirm, setShowConfirm] = useState(false);
+	const [isSubmitting, setIsSubmitting] = useState(false);
+	const { data, error, refetch } = useFetch(
 		`http://localhost:3030/data/comments/${comment._id}`,
 		null,
 		'DELETE',
@@ -15,12 +18,22 @@ export default function DeleteComment({ comment, updateCommentHandler }) {
 		false,
 	);
 
-	const deleteCommentHandler = async () => {
+	const handleCancelDelete = () => {
+		setShowConfirm(false);
+	};
+	const handleConfirmDelete = async () => {
+		setIsSubmitting(true);
 		await refetch(null, {
 			'X-Authorization': user?.accessToken,
 		});
-		updateCommentHandler({ type: 'delete', payload: comment._id });
+		setIsSubmitting(false);
 	};
+
+	useEffect(() => {
+		if (data) {
+			updateCommentHandler({ type: 'delete', payload: comment._id });
+		}
+	}, [updateCommentHandler, data, comment._id]);
 
 	if (error) {
 		//TODO add global error handling
@@ -32,7 +45,7 @@ export default function DeleteComment({ comment, updateCommentHandler }) {
 				type="button"
 				className={styles.deleteButton}
 				title="Delete comment"
-				onClick={() => deleteCommentHandler()}
+				onClick={() => setShowConfirm(true)}
 			>
 				<svg
 					width="18"
@@ -50,7 +63,15 @@ export default function DeleteComment({ comment, updateCommentHandler }) {
 					/>
 				</svg>
 			</button>
-			<DeleteConfirmationModal />
+			{showConfirm && (
+				<DeleteConfirmationModalPortal>
+					<DeleteConfirmationModal
+						handleCancelDelete={handleCancelDelete}
+						handleConfirmDelete={handleConfirmDelete}
+						isSubmitting={isSubmitting}
+					/>
+				</DeleteConfirmationModalPortal>
+			)}
 		</>
 	);
 }
